+++ resources/graphics.py
@@ -531,6 +531,8 @@
 				"search",
 				"searchOn",
 				"home",
+				"historyBack",
+				"historyForward",
 			],
 
 		},